requirements.txt
attr==0.3.2
colorama==0.4.6
ConfigParser==7.2.0
cryptography==45.0.4
docutils==0.21.2
filelock==3.18.0
gymnasium==1.1.1
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.7
jnius==1.1.0
joblib==1.4.2
keyring==25.6.0
matplotlib==3.8.4
numpy==2.3.1
optuna==4.1.0
pandas==2.3.0
Pillow==11.2.1
protobuf==6.31.1
pyOpenSSL==25.1.0
pytest==8.4.1
PyYAML==6.0.2
PyYAML==6.0.2
redis==6.2.0
scikit_learn==1.7.0
selenium==4.33.0
selenium_wire==5.1.0
setuptools==80.9.0
Sphinx==8.2.3
statsmodels==0.14.4
thread==2.0.5
torch==2.7.1
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
zstandard==0.22.0
